requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pytest==8.3.3
python-docx==1.2.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
